# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Duollance Client Acquisition Tool API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Lead {
  id: number;
  companyName: string;
  contactName: string;
  /** @nullable */
  contactTitle?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactLinkedIn?: string | null;
  /** @nullable */
  companySize?: string | null;
  /** @nullable */
  industry?: string | null;
  /** reddit, linkedin, twitter, producthunt, indiehackers, manual */
  source: string;
  /** @nullable */
  sourceUrl?: string | null;
  /**
     * The raw post or signal that surfaced this lead
     * @nullable
     */
  sourceContext?: string | null;
  /** cost, reliability, speed, vetting, other */
  painPoint: string;
  /** 1-100 Duollance fit score */
  fitScore: number;
  /** discovered, contacted, in_discussion, onboarded, rejected */
  stage: string;
  approved: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  followUpAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface LeadInput {
  companyName: string;
  contactName: string;
  contactTitle?: string;
  contactEmail?: string;
  contactLinkedIn?: string;
  companySize?: string;
  industry?: string;
  source: string;
  sourceUrl?: string;
  sourceContext?: string;
  painPoint: string;
  fitScore: number;
  notes?: string;
  followUpAt?: string;
}

export interface LeadUpdate {
  companyName?: string;
  contactName?: string;
  contactTitle?: string;
  contactEmail?: string;
  contactLinkedIn?: string;
  companySize?: string;
  industry?: string;
  source?: string;
  sourceUrl?: string;
  sourceContext?: string;
  painPoint?: string;
  fitScore?: number;
  stage?: string;
  approved?: boolean;
  notes?: string;
  followUpAt?: string;
}

export interface StageUpdate {
  stage: string;
}

export interface OutreachMessage {
  id: number;
  leadId: number;
  /** email, linkedin, whatsapp, twitter */
  channel: string;
  message: string;
  /** draft, sent, replied, no_response */
  status: string;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
}

export interface OutreachInput {
  leadId: number;
  channel: string;
  message: string;
  status?: string;
}

export interface OutreachUpdate {
  status?: string;
  message?: string;
  sentAt?: string;
}

export interface PitchRequest {
  leadId: number;
  channel?: string;
}

export interface PitchResponse {
  pitch: string;
}

export interface ScoreRequest {
  companyName: string;
  sourceContext: string;
  painPoint: string;
  industry?: string;
}

export interface ScoreResponse {
  score: number;
  reasoning: string;
  painPointCategory: string;
}

export interface DashboardStats {
  totalLeads: number;
  approvedLeads: number;
  contacted: number;
  inDiscussion: number;
  onboarded: number;
  avgFitScore: number;
  newToday: number;
}

export interface PipelineStageCount {
  stage: string;
  count: number;
}

export interface SourceCount {
  source: string;
  count: number;
}

export type ListLeadsParams = {
status?: string;
source?: string;
search?: string;
};

export type ListOutreachParams = {
leadId?: number;
};