# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Duollance Client Acquisition Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalLeads: number;
  approvedLeads: number;
  contacted: number;
  inDiscussion: number;
  onboarded: number;
  avgFitScore: number;
  newToday: number;
}