# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Duollance Client Acquisition Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface Lead {
  id: number;
  companyName: string;
  contactName: string;
  /** @nullable */
  contactTitle?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactLinkedIn?: string | null;
  /** @nullable */
  companySize?: string | null;
  /** @nullable */
  industry?: string | null;
  /** reddit, linkedin, twitter, producthunt, indiehackers, manual */
  source: string;
  /** @nullable */
  sourceUrl?: string | null;
  /**
     * The raw post or signal that surfaced this lead
     * @nullable
     */
  sourceContext?: string | null;
  /** cost, reliability, speed, vetting, other */
  painPoint: string;
  /** 1-100 Duollance fit score */
  fitScore: number;
  /** discovered, contacted, in_discussion, onboarded, rejected */
  stage: string;
  approved: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  followUpAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}