# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Duollance Client Acquisition Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadInput {
  companyName: string;
  contactName: string;
  contactTitle?: string;
  contactEmail?: string;
  contactLinkedIn?: string;
  companySize?: string;
  industry?: string;
  source: string;
  sourceUrl?: string;
  sourceContext?: string;
  painPoint: string;
  fitScore: number;
  notes?: string;
  followUpAt?: string;
}