# lib/api-zod/src/generated/types/outreachMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Duollance Client Acquisition Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface OutreachMessage {
  id: number;
  leadId: number;
  /** email, linkedin, whatsapp, twitter */
  channel: string;
  message: string;
  /** draft, sent, replied, no_response */
  status: string;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
}