# lib/api-zod/src/generated/types/pipelineStageCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Duollance Client Acquisition Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface PipelineStageCount {
  stage: string;
  count: number;
}